Add files via upload
In this exercise, we wish to understand why the wafers we are using have a thickness of 220 nm.

You can use the analytic method with the script provided below (in MATLAB) to answer this question (Note: you do not need to understand the details of this script, just use it as a tool. But, you will need to change the input paramaters based on the calculation you want to do.).

You can download the script (wg_1D_slab.m).

What is the maximum Si thickness before the slab waveguide supports more than one TE mode? Enter the thickness [Units: nm, accuracy +/-5 nm]: